+++ include/dynd/exceptions.hpp
@@ -74,7 +74,7 @@ public:
      * An exception for when too many indices are provided in
      * an indexing operation (nindex > ndim).
      */
-    too_many_indices(int nindices, int ndim);
+    too_many_indices(size_t nindices, size_t ndim);
 
     virtual ~too_many_indices() throw() {
     }
@@ -86,8 +86,8 @@ public:
      * An exception for when 'i' isn't within bounds for
      * the specified axis of the given shape
      */
-    index_out_of_bounds(intptr_t i, int axis, int ndim, const intptr_t *shape);
-    index_out_of_bounds(intptr_t i, int axis, const std::vector<intptr_t>& shape);
+    index_out_of_bounds(intptr_t i, size_t axis, size_t ndim, const intptr_t *shape);
+    index_out_of_bounds(intptr_t i, size_t axis, const std::vector<intptr_t>& shape);
     index_out_of_bounds(intptr_t i, intptr_t dimension_size);
 
     virtual ~index_out_of_bounds() throw() {
@@ -100,7 +100,7 @@ public:
      * An exception for when 'i' isn't a valid axis
      * for the number of dimensions.
      */
-    axis_out_of_bounds(intptr_t i, intptr_t ndim);
+    axis_out_of_bounds(size_t i, size_t ndim);
 
     virtual ~axis_out_of_bounds() throw() {
     }
@@ -115,8 +115,8 @@ public:
      * An exception for when 'i' isn't within bounds for
      * the specified axis of the given shape
      */
-    irange_out_of_bounds(const irange& i, int axis, int ndim, const intptr_t *shape);
-    irange_out_of_bounds(const irange& i, int axis, const std::vector<intptr_t>& shape);
+    irange_out_of_bounds(const irange& i, size_t axis, size_t ndim, const intptr_t *shape);
+    irange_out_of_bounds(const irange& i, size_t axis, const std::vector<intptr_t>& shape);
     irange_out_of_bounds(const irange& i, intptr_t dimension_size);
 
     virtual ~irange_out_of_bounds() throw() {
