+++ src/dynd/dtypes/bytes_dtype.cpp
@@ -74,7 +74,7 @@ dtype bytes_dtype::get_canonical_dtype() const
 }
 
 
-void bytes_dtype::get_shape(int DYND_UNUSED(i), std::vector<intptr_t>& DYND_UNUSED(out_shape)) const
+void bytes_dtype::get_shape(size_t DYND_UNUSED(i), std::vector<intptr_t>& DYND_UNUSED(out_shape)) const
 {
 }
