+++ src/dynd/dtypes/byteswap_dtype.cpp
@@ -75,7 +75,7 @@ dtype dynd::byteswap_dtype::apply_linear_index(int nindices, const irange *indic
     }
 }
 
-void dynd::byteswap_dtype::get_shape(int i, intptr_t *out_shape) const
+void dynd::byteswap_dtype::get_shape(size_t i, intptr_t *out_shape) const
 {
     if (!m_operand_dtype.is_builtin()) {
         m_operand_dtype.extended()->get_shape(i, out_shape);
