+++ src/dynd/dtypes/categorical_dtype.cpp
@@ -251,7 +251,7 @@ dtype dynd::categorical_dtype::apply_linear_index(int nindices, const irange *in
     }
 }
 
-void dynd::categorical_dtype::get_shape(int i, intptr_t *out_shape) const
+void dynd::categorical_dtype::get_shape(size_t i, intptr_t *out_shape) const
 {
     if (!m_category_dtype.is_builtin()) {
         m_category_dtype.extended()->get_shape(i, out_shape);
